*fix: infinite analyze phase when reprocessing assets with "skip startup scan" is enable.
After AP launch when assets were modified or asked to reporcess, AP stayed in analyze phase indefinitely.
The problem was due to missing setting of m_isCurrentlyScanning to false before requier Idel phase.
For this reason reporcess stay queued and the system enters in an infinite loop between QueueIdleCheck and ScheduleNextUpdate calls